reducing API calls by doing it all at first

+++ src/components/Session.js
@@ -10,7 +10,6 @@ class Session extends Component {
             date: '',
             sessionLength: '',
             location: '',
-            ID: props.sessionInfo.ID,
             studentConfirmed: props.sessionInfo.StudentConfirmed,
             tutorConfirmed: props.sessionInfo.TutorConfirmed
         }
@@ -19,33 +18,35 @@ class Session extends Component {
     }
     
     handleCancel(){
-        let {ID} = this.props.sessionInfo
-        alert('Cancelled' + this.props.sessionInfo.ID)
-        let endpoint = "https://y9ynb3h6ik.execute-api.us-east-1.amazonaws.com/prodAPI/sessions?ID=" + ID
-        fetch(endpoint, {method: "DELETE"})
-        .then(response => response.json())
-        .then(response => {
-            alert('Session was successfully cancelled')
-            this.setState({
-                show: false
+        if (confirm("Are you sure you want to cancel this session?")) {
+            let {ID} = this.props.sessionInfo
+            let endpoint = "https://y9ynb3h6ik.execute-api.us-east-1.amazonaws.com/prodAPI/sessions?ID=" + ID
+            fetch(endpoint, {method: "DELETE"})
+            .then(response => response.json())
+            .then(response => {
+                alert('Session was successfully cancelled')
+                this.setState({
+                    show: false
+                })
             })
-        })
-        .catch(err => alert("Error cancelling session: " + err))
+            .catch(err => alert("Error cancelling session: " + err))
+        }
     }
     
     confirm(){
         let endpoint = "https://y9ynb3h6ik.execute-api.us-east-1.amazonaws.com/prodAPI/sessions"
+        let {ID} = this.props.sessionInfo
 
         if (this.props.userRole === 'Student') {
             this.setState({
                 studentConfirmed: true
             })
-            endpoint += '?userRole=Student&sessionID=' + this.state.ID
+            endpoint += '?userRole=Student&sessionID=' + ID
         } else if (this.props.userRole === 'Tutor') {
             this.setState({
                 tutorConfirmed: true
             })
-            endpoint += '?userRole=Tutor&sessionID=' + this.state.ID
+            endpoint += '?userRole=Tutor&sessionID=' + ID
         }
         fetch(endpoint, {method: "PUT"})
         .catch(err => console.log("ERR: " + err))
@@ -54,22 +55,22 @@ class Session extends Component {
     render(){
         let {Tutor, Student, Subject, Location, date} = this.props.sessionInfo
         let {studentConfirmed, tutorConfirmed, show} = this.state
-        let {userRole} = this.props
+        let {userRole, secondaryRole} = this.props
         let dateFormatted = new Date(Date.parse(date))
         return (
         <div className = {show ? "sessionWrapper" : "hiddenWrapper"}>
             {userRole !== 'Student' ? <div>Student: {Student} </div> : null}
             {userRole !== 'Tutor' ? <div> Tutor: {Tutor} </div> : null}
             <div> Subject: {Subject}
-            {userRole == 'Student' || userRole == 'Tutor' ? <button className="cancel-button" onClick = {this.handleCancel}> Cancel </button> : null}
+            {(userRole == 'Student' || userRole == 'Tutor') && secondaryRole !== 'Admin' && new Date() < new Date(Date.parse(date)) ? <button className="cancel-button" onClick = {this.handleCancel}> Cancel </button> : null}
             </div>
             <div>
-            Location: {Location} on {dateFormatted.toLocaleDateString()} at  {dateFormatted.toLocaleTimeString()}
+            Location: {Location} on {dateFormatted.toLocaleDateString()} at {dateFormatted.toLocaleTimeString()}
             </div>
             {userRole == 'Student' && !studentConfirmed && 
-            !this.props.secondaryRole? // && this.props.isPrimary ? 
+            !this.props.secondaryRole && new Date() < new Date(Date.parse(date)) ? // && this.props.isPrimary ? 
                 <button onClick={this.confirm}> Confirm </button> : null}
-            {userRole == 'Tutor' && !tutorConfirmed && !this.props.secondaryRole? //&& this.props.isPrimary ?
+            {userRole == 'Tutor' && !tutorConfirmed && !this.props.secondaryRole && new Date() < new Date(Date.parse(date)) ? //&& this.props.isPrimary ?
                 <button onClick={this.confirm}> Confirm </button> : null}
             <div>
                 StudentConfirmed: {studentConfirmed == 1? 'yes' : 'no'}
